changed up the components DS
